requirements.txt
requests
langchain-openai
python-dotenv
psycopg2-binary
pgvector
